requirements.txt
torchvision==0.10.0
tensorflow-gpu==2.5.0
tensorflow_datasets==4.3.0

pandas==1.2.4
numpy==1.19.5
sklearn==0.24.2
nltk==3.6.2
transformers==4.6.1
ijson==3.1.4
sty==1.0.0
matplotlib==3.4.2

fvcore==0.1.5.post20210617

elasticsearch==7.13.2
